Implementação de algoritmos de ordenação

+++ Sorting-Algorithms/src/br/facisa/p3/ed/sortingAlgorithms/BubbleSort.java
@@ -10,11 +10,11 @@ public class BubbleSort {
 	 * @param array
 	 * @return sorted array
 	 */
-	public static double[] sort (double[] array){
+	public static int[] sort (int[] array){
 		for (int i = array.length;  i > 0; i--){
 			for (int j = 0; j < i-1; j++){
 				if (array[j] > array[j+1]){
-					double temp = array[j+1];
+					int temp = array[j+1];
 					array[j+1] = array[j];
 					array[j] = temp;
 				}
